A minimalist browser
The code is taken from https://github.com/sapgan/browser and this provides a layout of a browser starting from "http://www.google.com". Most of the fancy animations and images are missing from this browser, however it can be used to connect to the internet and perform some tasks. The task was due on 27th January 2019, (Sunday Fun)